+++ core/src/main/java/org/eqasim/core/components/transit/routing/DefaultEnrichedTransitRoute.java
@@ -5,9 +5,12 @@ import java.io.IOException;
 import org.matsim.api.core.v01.Id;
 import org.matsim.api.core.v01.network.Link;
 import org.matsim.core.population.routes.AbstractRoute;
+import org.matsim.core.utils.misc.OptionalTime;
+import org.matsim.pt.routes.DefaultTransitPassengerRoute.RouteDescription;
 import org.matsim.pt.transitSchedule.api.Departure;
 import org.matsim.pt.transitSchedule.api.TransitLine;
 import org.matsim.pt.transitSchedule.api.TransitRoute;
+import org.matsim.pt.transitSchedule.api.TransitStopFacility;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.core.JsonProcessingException;
@@ -16,7 +19,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 public class DefaultEnrichedTransitRoute extends AbstractRoute implements EnrichedTransitRoute {
 	final public static String ROUTE_TYPE = "enriched_pt";
 
-	private RouteDescription routeDescription = null;
+	private EnrichedRouteDescription routeDescription = null;
 
 	public DefaultEnrichedTransitRoute(final Id<Link> startLinkId, final Id<Link> endLinkId) {
 		super(startLinkId, endLinkId);
@@ -24,13 +27,15 @@ public class DefaultEnrichedTransitRoute extends AbstractRoute implements Enrich
 
 	public DefaultEnrichedTransitRoute(final Id<Link> startLinkId, final Id<Link> endLinkId, double distance,
 			double inVehicleTime, double transferTime, int accessStopIndex, int egressStopIndex,
-			Id<TransitLine> transitLineId, Id<TransitRoute> transitRouteId, Id<Departure> departureId) {
+			Id<TransitLine> transitLineId, Id<TransitRoute> transitRouteId, Id<Departure> departureId,
+			Id<TransitStopFacility> accessFacilityId, Id<TransitStopFacility> egressFacilityId,
+			OptionalTime boardingTime) {
 		super(startLinkId, endLinkId);
 
 		setDistance(distance);
 		setTravelTime(inVehicleTime + transferTime);
 
-		this.routeDescription = new RouteDescription();
+		this.routeDescription = new EnrichedRouteDescription();
 		routeDescription.inVehicleTime = inVehicleTime;
 		routeDescription.transferTime = transferTime;
 		routeDescription.accessStopIndex = accessStopIndex;
@@ -38,6 +43,9 @@ public class DefaultEnrichedTransitRoute extends AbstractRoute implements Enrich
 		routeDescription.transitLineId = transitLineId;
 		routeDescription.transitRouteId = transitRouteId;
 		routeDescription.departureId = departureId;
+		routeDescription.accessFacilityId = accessFacilityId;
+		routeDescription.egressFacilityId = egressFacilityId;
+		routeDescription.boardingTime = boardingTime;
 	}
 
 	@Override
@@ -57,7 +65,7 @@ public class DefaultEnrichedTransitRoute extends AbstractRoute implements Enrich
 	@Override
 	public void setRouteDescription(String routeDescription) {
 		try {
-			this.routeDescription = new ObjectMapper().readValue(routeDescription, RouteDescription.class);
+			this.routeDescription = new ObjectMapper().readValue(routeDescription, EnrichedRouteDescription.class);
 		} catch (IOException e) {
 			throw new RuntimeException(e);
 		}
@@ -74,12 +82,12 @@ public class DefaultEnrichedTransitRoute extends AbstractRoute implements Enrich
 	}
 
 	@Override
-	public Id<TransitLine> getTransitLineId() {
+	public Id<TransitLine> getLineId() {
 		return routeDescription.transitLineId;
 	}
 
 	@Override
-	public Id<TransitRoute> getTransitRouteId() {
+	public Id<TransitRoute> getRouteId() {
 		return routeDescription.transitRouteId;
 	}
 
@@ -101,49 +109,42 @@ public class DefaultEnrichedTransitRoute extends AbstractRoute implements Enrich
 	@Override
 	public DefaultEnrichedTransitRoute clone() {
 		return new DefaultEnrichedTransitRoute(getStartLinkId(), getEndLinkId(), getDistance(), getInVehicleTime(),
-				getWaitingTime(), getAccessStopIndex(), getEgressStopIndex(), getTransitLineId(), getTransitRouteId(),
-				getDepartureId());
+				getWaitingTime(), getAccessStopIndex(), getEgressStopIndex(), getLineId(), getRouteId(),
+				getDepartureId(), getAccessStopId(), getEgressStopId(), getBoardingTime());
 	}
 
-	public static class RouteDescription {
+	public static class EnrichedRouteDescription extends RouteDescription {
 		public double inVehicleTime;
 		public double transferTime;
 
 		public int accessStopIndex;
 		public int egressStopindex;
 
-		public Id<TransitLine> transitLineId;
-		public Id<TransitRoute> transitRouteId;
 		public Id<Departure> departureId;
-
-		@JsonProperty("transitLineId")
-		public String getTransitLineId() {
-			return transitLineId.toString();
-		}
-
-		@JsonProperty("transitRouteId")
-		public String getRouteLineId() {
-			return transitRouteId.toString();
-		}
 
 		@JsonProperty("departureId")
 		public String getDepartureId() {
 			return departureId.toString();
 		}
 
-		@JsonProperty("transitLineId")
-		public void setTransitLineId(String transitLineId) {
-			this.transitLineId = Id.create(transitLineId, TransitLine.class);
-		}
-
-		@JsonProperty("transitRouteId")
-		public void setRouteLineId(String transitRouteId) {
-			this.transitRouteId = Id.create(transitRouteId, TransitRoute.class);
-		}
-
 		@JsonProperty("departureId")
 		public void setDepartureId(String departureId) {
 			this.departureId = Id.create(departureId, Departure.class);
 		}
+	}
+
+	@Override
+	public Id<TransitStopFacility> getAccessStopId() {
+		return routeDescription.accessFacilityId;
+	}
+
+	@Override
+	public Id<TransitStopFacility> getEgressStopId() {
+		return routeDescription.egressFacilityId;
+	}
+
+	@Override
+	public OptionalTime getBoardingTime() {
+		return routeDescription.boardingTime;
 	}
 }